Move EL related tests to cdi-full group, #340

Document the cdiLiteMode flag and add a core profile config example, … (#339)

* Document the cdiLiteMode flag and add a core profile config example, #335
* Add a ref to cdiLiteMode in getValue IllegalArgumentException.


cdi.api.version to 4.0.0-RC4

[maven-release-plugin] prepare release 4.0.0-RC4

[maven-release-plugin] prepare for next development iteration

Move EL related tests to cdi-full group, #340
